AP-279 refactor PluginType SourceType namespace
